+++ Game/src/sys/CameraSystem.cpp
@@ -2,14 +2,14 @@
 #include<comp/actorComponents.h>
 #include <util/isomath.h>
 
-CameraSystem::CameraSystem(int x, int y, entt::registry& registry) : System(registry), mBaseXOffset(x), mBaseYOffset(y){}
+CameraSystem::CameraSystem(float x, float y, entt::registry& registry) : System(registry), mBaseXOffset(x), mBaseYOffset(y){}
 
 void CameraSystem::update(double dt) {
 	mRegistry.view<camera, position>().each([&](auto& camera, auto& position) {
 		camera.xOffset = -position.x;
 		camera.yOffset = -position.y;
 		twoDToIso(camera.xOffset, camera.yOffset);
-		camera.xOffset = camera.xOffset +mBaseXOffset;
-		camera.yOffset = camera.yOffset +mBaseYOffset;
+		camera.xOffset = camera.xOffset + mBaseXOffset;
+		camera.yOffset = camera.yOffset + mBaseYOffset;
 	});
 }
